+++ client/js/components/procedure/admin/AdministrationCustomFieldsList.vue
@@ -6,41 +6,48 @@
       dismissible
       :dismissible-key="helpTextDismissibleKey"
       :message="Translator.trans('segments.fields.edit.info')"
-      type="info" />
+      type="info"
+    />
 
     <create-custom-field-form
       :handle-success="isSuccess"
       :is-loading="isLoading"
-      @save="customFieldData => saveNewField(customFieldData)">
+      @save="customFieldData => saveNewField(customFieldData)"
+    >
       <div>
         <dp-label
           class="mb-1"
           required
-          :text="Translator.trans('options')" />
+          :text="Translator.trans('options')"
+        />
         <dp-input
           id="newFieldOption:1"
           v-model="newFieldOptions[0].label"
           class="mb-2 w-[calc(100%-26px)]"
           data-cy="customFields:newFieldOption1"
           maxlength="250"
-          required />
+          required
+        />
         <dp-input
           id="newFieldOption:2"
           v-model="newFieldOptions[1].label"
           class="mb-2 w-[calc(100%-26px)]"
           data-cy="customFields:newFieldOption2"
           maxlength="250"
-          required />
+          required
+        />
 
         <div
           v-for="(option, idx) in additionalOptions"
-          :key="`option:${idx}`">
+          :key="`option:${idx}`"
+        >
           <div class="w-[calc(100%-26px)] inline-block mb-2">
             <dp-input
               :id="`option:${newFieldOptions[idx + 2].label}`"
               v-model="newFieldOptions[idx + 2].label"
               :data-cy="`customFields:newFieldOption${idx + 2}`"
-              maxlength="250" />
+              maxlength="250"
+            />
           </div>
           <dp-button
             :data-cy="`customFields:removeOptionInput:${option.label}`"
@@ -49,7 +56,8 @@
             hide-text
             icon="x"
             variant="subtle"
-            @click="removeOptionInput(idx + 2)" />
+            @click="removeOptionInput(idx + 2)"
+          />
         </div>
 
         <dp-button
@@ -57,7 +65,8 @@
           data-cy="customFields:addOption"
           icon="plus"
           variant="subtle"
-          @click="addOptionInput" />
+          @click="addOptionInput"
+        />
       </div>
     </create-custom-field-form>
 
@@ -68,7 +77,8 @@
       data-cy="customFields:table"
       data-dp-validate="editCustomFieldsForm"
       has-flyout
-      track-by="id">
+      track-by="id"
+    >
       <template v-slot:name="rowData">
         <div v-if="rowData.edit">
           <dp-input
@@ -88,7 +98,8 @@
             v-for="(option, index) in displayedOptions(rowData)"
             :key="index"
             :data-cy="`customFields:option${option.label}`"
-            class="mb-1">
+            class="mb-1"
+          >
             <div>
               {{ option.label }}
             </div>
@@ -98,7 +109,8 @@
           <li
             v-for="(option, index) in newRowData.options"
             :key="index"
-            class="mb-1">
+            class="mb-1"
+          >
             <div class="flex">
               <dp-input
                 :id="`option:${index}`"
@@ -124,7 +136,8 @@
               icon="plus"
               variant="subtle"
               :text="Translator.trans('option.add')"
-              @click="addOptionInputOnEdit(rowData)" />
+              @click="addOptionInputOnEdit(rowData)"
+            />
           </li>
         </ul>
       </template>
@@ -133,7 +146,8 @@
         <div v-if="rowData.edit">
           <dp-input
             id="customFieldDescription"
-            v-model="newRowData.description" />
+            v-model="newRowData.description"
+          />
         </div>
         <div v-else>
           {{ rowData.description }}
@@ -148,7 +162,8 @@
             data-cy="customFields:editField"
             :aria-label="Translator.trans('item.edit')"
             :title="Translator.trans('edit')"
-            @click="editCustomField(rowData)">
+            @click="editCustomField(rowData)"
+          >
             <dp-icon
               aria-hidden="true"
               icon="edit"
@@ -182,7 +197,8 @@
               :title="Translator.trans('save')"
               class="btn--blank o-link--default u-mr-0_25 inline-block"
               data-cy="customFields:saveEdit"
-              @click="dpValidateAction('editCustomFieldsForm', () => saveEditedFields(), false)">
+              @click="dpValidateAction('editCustomFieldsForm', () => saveEditedFields(), false)"
+            >
               <dp-icon
                 aria-hidden="true"
                 icon="check"
@@ -194,7 +210,8 @@
               :title="Translator.trans('abort')"
               class="btn--blank o-link--default inline-block"
               data-cy="customFields:abortEdit"
-              @click="abortFieldEdit(rowData)">
+              @click="abortFieldEdit(rowData)"
+            >
               <dp-icon
                 aria-hidden="true"
                 icon="xmark"
@@ -214,10 +231,12 @@
             :disabled="rowData.options.length < 3"
             class="btn--blank o-link--default"
             data-cy="customFields:showOptions"
-            @click="showOptions(rowData)">
+            @click="showOptions(rowData)"
+          >
             <dp-icon
               aria-hidden="true"
-              icon="caret-down" />
+              icon="caret-down"
+            />
           </button>
 
           <template v-else>
@@ -226,10 +245,12 @@
               :aria-label="Translator.trans('aria.collapse')"
               class="btn--blank o-link--default"
               data-cy="customFields:hideOptions"
-              @click="hideOptions(rowData)">
+              @click="hideOptions(rowData)"
+            >
               <dp-icon
                 aria-hidden="true"
-                icon="caret-up" />
+                icon="caret-up"
+              />
             </button>
           </template>
         </div>