+++ client/js/components/statement/StatementExportModal.vue
@@ -13,12 +13,14 @@
       data-cy="exportModal:open"
       :text="Translator.trans('export.verb')"
       variant="subtle"
-      @click.prevent="openModal" />
+      @click.prevent="openModal"
+    />
 
     <dp-modal
       ref="exportModalInner"
       content-classes="w-11/12 sm:w-10/12 md:w-8/12 lg:w-6/12 xl:w-5/12 h-fit"
-      content-body-classes="flex flex-col h-[95%]">
+      content-body-classes="flex flex-col h-[95%]"
+    >
       <h2 class="mb-5">
         {{ exportModalTitle }}
       </h2>
@@ -26,7 +28,8 @@
       <fieldset v-if="!isSingleStatementExport">
         <legend
           class="o-form__label text-base"
-          v-text="Translator.trans('export.type')" />
+          v-text="Translator.trans('export.type')"
+        />
         <div class="grid grid-cols-3 mt-2 mb-5 gap-x-2 gap-y-5">
           <dp-radio
             v-for="(exportType, key) in exportTypes"
@@ -39,26 +42,30 @@
             }"
             :value="key"
             :checked="active === key"
-            @change="active = key" />
+            @change="active = key"
+          />
           <template v-if="active !== 'xlsx_normal'">
             <dp-checkbox
               id="censoredCitizen"
               v-model="isCitizenDataCensored"
               :label="{
                 text: Translator.trans('export.censored.citizen')
-              }" />
+              }"
+            />
             <dp-checkbox
               id="censoredInstitution"
               v-model="isInstitutionDataCensored"
               :label="{
                 text: Translator.trans('export.censored.institution')
-              }" />
+              }"
+            />
             <dp-checkbox
               id="obscured"
               v-model="isObscure"
               :label="{
                 text: Translator.trans('export.docx.obscured')
-              }" />
+              }"
+            />
           </template>
         </div>
       </fieldset>
@@ -70,19 +77,22 @@
             v-model="isCitizenDataCensored"
             :label="{
               text: Translator.trans('export.censored.citizen')
-            }" />
+            }"
+          />
           <dp-checkbox
             id="singleStatementInstitution"
             v-model="isInstitutionDataCensored"
             :label="{
               text: Translator.trans('export.censored.institution')
-            }" />
+            }"
+          />
           <dp-checkbox
             id="singleStatementObscure"
             v-model="isObscure"
             :label="{
               text: Translator.trans('export.docx.obscured')
-            }" />
+            }"
+          />
         </div>
       </fieldset>
 
@@ -90,10 +100,12 @@
         <legend
           id="docxColumnTitles"
           class="o-form__label text-base float-left mr-1"
-          v-text="Translator.trans('docx.export.column.title')" />
+          v-text="Translator.trans('docx.export.column.title')"
+        />
         <dp-contextual-help
           aria-labelledby="docxColumnTitles"
-          :text="Translator.trans('docx.export.column.title.hint')" />
+          :text="Translator.trans('docx.export.column.title.hint')"
+        />
         <div class="grid grid-cols-5 gap-3 mt-1 mb-5">
           <dp-input
             v-for="(column, key) in docxColumns"
@@ -103,26 +115,31 @@
             :data-cy="column.dataCy"
             :placeholder="Translator.trans(column.placeholder)"
             type="text"
-            :width="column.width" />
+            :width="column.width"
+          />
         </div>
         <fieldset v-if="active === 'zip' || isSingleStatementExport">
           <legend
             id="docxFileName"
             class="o-form__label text-base float-left mr-1"
-            v-text="Translator.trans('docx.export.file_name')" />
+            v-text="Translator.trans('docx.export.file_name')"
+          />
           <dp-contextual-help
             aria-labelledby="docxFileName"
-            :text="Translator.trans('docx.export.file_name.hint')" />
+            :text="Translator.trans('docx.export.file_name.hint')"
+          />
           <dp-input
             id="fileName"
             v-model="fileName"
             class="mt-1"
             :placeholder="Translator.trans('docx.export.file_name.placeholder')"
-            type="text" />
+            type="text"
+          />
           <div class="font-size-small mt-2">
             <span
               class="weight--bold"
-              v-text="Translator.trans('docx.export.example_file_name')" />
+              v-text="Translator.trans('docx.export.example_file_name')"
+            />
             <span v-text="exampleFileName" />
           </div>
         </fieldset>
@@ -136,7 +153,8 @@
         :primary-text="Translator.trans('export.statements')"
         :secondary-text="Translator.trans('abort')"
         @primary-action="handleExport"
-        @secondary-action="closeModal" />
+        @secondary-action="closeModal"
+      />
     </dp-modal>
   </div>
 </template>