+++ client/js/components/user/DpTableCardList/DpTableCardListHeader.vue
@@ -6,7 +6,8 @@
       class="layout__item u-1-of-2 u-mb-0_5"
       :placeholder="searchPlaceholder"
       @reset="$emit('resetSearch')"
-      @search="val => $emit('search', val)" /><!--
+      @search="val => $emit('search', val)"
+    /><!--
  --><slot name="header-buttons" />
     <!-- header with checkbox and labels-->
     <div class="layout__item border--bottom">
@@ -14,12 +15,14 @@
         v-if="selectable"
         id="selectAll"
         class="inline-block w-[20px] u-pv-0_25"
-        @change="val => $emit('selectAll', val)" /><!--
+        @change="val => $emit('selectAll', val)"
+      /><!--
     --><div
         v-for="(item, idx) in items"
         :key="idx"
         class="layout__item weight--bold u-pv-0_5"
-        :class="[item.classes ? item.classes : '', item.width ? item.width : '']">
+        :class="[item.classes ? item.classes : '', item.width ? item.width : '']"
+      >
         {{ item.label }}
       </div>
     </div>