+++ CHANGELOG.md
@@ -5,6 +5,23 @@
 - **Patch Version**: Incremented for bug fixes.
 
 ## UNRELEASED
+- Allow project specific CSS
+- allow sessions to be stored in redis
+
+-  Add checkbox in procedure settings to expand procedure description in public view on page load
+- Use external Geocoder API as service for address auto-suggestions
+
+- Turn projects into yarn workspaces
+
+## v4.15.0 (2025-09-15)
+## v4.14.0 (2025-09-15)
+- Add html paragraph import from odt files
+
+## v4.12.0 (2025-09-10)
+- Add ODT export functionality for assessment tables
+- Add checkbox in procedure settings to expand procedure description in public view on page load
+- Use external Geocoder API as service for address auto-suggestions
+- Update demosplan-addon from v0.59 to v0.60
 
 ## v4.11.0 (2025-08-27)
 - Allow to edit custom field type singleSelect